代码优化

+++ server/src/main/java/com/atey/task/OrderTask.java
@@ -81,7 +81,9 @@ public class OrderTask {
                 idList.add(marketProducesPlus.getId());
             }
         }
-        marketProducesPlusMapper.deleteByIds(idList);
+        if(!idList.isEmpty()){
+            marketProducesPlusMapper.deleteByIds(idList);
+        }
 
         List<MarketProduces> list = Db.lambdaQuery(MarketProduces.class)
                 .eq(MarketProduces::getDeleted, DeletedConstant.NOT_DELETED)
